requirements.txt
streamlit
sqlite3
matplotlib
plotly
streamlit-authenticator
passlib
pandas
schedule
pyyaml
smtplib
